Fix errors with the tileLayer.
Added a --env=dev option to the Gruntfile to build with source maps.

Fixed an issue in the camera where the translation was not being scaled.  mat4.scale doesn't affect mat4.translate, so the translation had to be modified.

Fixed an issue in the map the set the left and right bounds using phiMax rather than +/- 180 degrees.

Added a map.worldToDisplay() method.

Added a concept of tileOffset.  Perhaps this should be done differently.

Fixed some comment typos (e.g., in modulo)

Fixed the css transform.

Simplified _getViewBounds, _toLocal, and _fromLocal.  I may have missed some of the original intent and should have done this differently.

Added a tileRounding option so that we can choose to have tiles that are scaled down rather than scaled up.

We should clamp panning if the bounds are clamped.